Add new type of airlock
This is meant to be used in hangars, where you have one giant room
that's constantly pressurized/depressurized, as opposed to a more
traditional decompression chamber airlock JAMS was originally designed
for. This takes any number of doors, sensors and air vents.

Additionally, multiple air vents are now possible to use with all
airlocks.